refactor: remove Title component and use HTML tag instead
