+++ src/app/sign-up/auth/page.tsx
@@ -1,10 +1,7 @@
-"use client";
-
 import { css } from "@/styled-system/css";
 import Link from "next/link";
 import React from "react";
 import Step from "@/src/components/common/Step";
-import Title from "@/src/components/common/Title";
 import AuthSection from "@/src/components/common/AuthSection";
 import HelpBox from "@/src/components/sign-up/HelpBox";
 
@@ -16,7 +13,14 @@ function page() {
         <Step.Connector />
         <Step.Circle stepNumber={2} text="소셜 계정 연동" />
       </Step.Root>
-      <Title />
+
+      <h1 className={css({ ...mainTextStyles })}>SOPT 회원인증</h1>
+      <p className={css({ ...textStyles, ...descriptionStyles })}>
+        이곳은 SOPT 회원만을 위한 공간이에요.
+        <br />
+        SOPT 회원인증을 위해 전화번호를 입력해 주세요.
+      </p>
+
       <AuthSection>
         <Link
           href="https://docs.google.com/forms/d/e/1FAIpQLSdBxksqlkAHShYdQYxDIK1Mnsy45MbYMkEeGuCMpeXjn6C1NQ/viewform"
@@ -43,4 +47,32 @@ const mainWrapperStyles = css.raw({
 
 const linkStyles = css.raw({
   marginTop: "4.1rem",
+});
+
+const mainTextStyles = css.raw({
+  textAlign: "center",
+  textStyle: "heading-2-32-b",
+  color: "white",
+  marginTop: "9.5rem",
+
+  "@media (max-width: 480px)": {
+    textStyle: "title-3-24-sb",
+    marginTop: "5.4rem",
+  },
+});
+
+const textStyles = css.raw({
+  display: "flex",
+  alignItems: "center",
+  textStyle: "body-2-16-m",
+  color: "gray.200",
+
+  "@media (max-width:480px)": {
+    textStyle: "label-4-12-sb",
+  },
+});
+
+const descriptionStyles = css.raw({
+  textAlign: "center",
+  margin: "1.4rem auto 0 auto",
 });
